+++ src/Molecules/Drawer/Drawer.tsx
@@ -3,26 +3,17 @@ import { useUIDSeed } from 'react-uid';
 import styled from 'styled-components';
 import FocusLock from 'react-focus-lock';
 import { RemoveScroll } from 'react-remove-scroll';
-import { motion } from 'framer-motion';
-import { DrawerComponent, TitleProps } from './Drawer.types';
+import { motion, useDragControls, AnimatePresence, PanInfo } from 'framer-motion';
+import { Props, TitleProps } from './Drawer.types';
 import { isBoolean, isElement } from '../../common/utils';
-import {
-  Flexbox,
-  Typography,
-  Icon,
-  Box,
-  useKeyPress,
-  FadedScroll,
-  Portal,
-  useMedia,
-  Button,
-} from '../..';
+import { Typography, Icon, useKeyPress, Portal, useMedia, Button } from '../..';
 
+const CROSS_SIZE = 5;
+const PADDING = 5;
 const displayName = 'Drawer';
 
 const Container = styled(motion.div)`
   box-sizing: border-box;
-  padding: ${({ theme }) => theme.spacing.unit(5)}px;
   width: 100%;
   height: 100%;
   background: ${({ theme }) => theme.color.card};
@@ -44,37 +35,53 @@ const Container = styled(motion.div)`
 `;
 
 const CloseButton = styled(Button)`
-  margin-left: auto;
+  position: absolute;
+  top: ${p => p.theme.spacing.unit(PADDING)}px;
+  right: ${p => p.theme.spacing.unit(PADDING)}px;
 `;
 
 const Content = styled.div`
   overflow-y: auto;
   overflow-x: hidden;
+  margin-bottom: ${p => p.theme.spacing.unit(PADDING)}px;
+  padding: 0 ${p => p.theme.spacing.unit(PADDING)}px;
 `;
 
 const H2 = styled.h2`
   padding-right: ${p => p.theme.spacing.unit(4)}px;
 `;
 
+const TitleWrapper = styled.div`
+  position: relative;
+  padding: ${p =>
+    `${p.theme.spacing.unit(PADDING)}px ${p.theme.spacing.unit(PADDING)}px 0 ${p.theme.spacing.unit(
+      PADDING,
+    )}px`};
+  margin-bottom: ${p => p.theme.spacing.unit(2)}px;
+  min-height: ${p => p.theme.spacing.unit(CROSS_SIZE)}px;
+  flex: 0 0 auto;
+`;
+
 const animationProps = {
-  initial: {
-    opacity: 0,
-    x: 70,
-  },
   animate: {
     opacity: 1,
     x: 0,
   },
   transition: {
-    type: 'spring',
-    damping: 15,
-    stiffness: 200,
+    ease: 'easeInOut',
+    duration: 0.2,
   },
 };
 
+// Todo: handle rollback if small offset
+const shouldCloseBecauseOfDrag = (info: PanInfo) =>
+  Math.abs(info.offset.x) > 0 || Math.abs(info.velocity.x) > 300;
+
 const components = {
+  CloseButton,
   Container,
   Content,
+  TitleWrapper,
 };
 
 const Title: React.FC<TitleProps> = ({ title, uid }) => {
@@ -91,50 +98,120 @@ const Title: React.FC<TitleProps> = ({ title, uid }) => {
   );
 };
 
-export const Drawer: DrawerComponent & {
+const useDifferentAnimationBasedOnSwipeDirection = () => {
+  const [swipeDirection, setSwipeDirection] = React.useState('right');
+  const exitAnimation = React.useMemo(
+    () => ({
+      opacity: 0,
+      x: swipeDirection === 'right' ? 700 : -700,
+    }),
+    [swipeDirection],
+  );
+  const initialAnimation = React.useMemo(
+    () => ({
+      opacity: 0,
+      x: swipeDirection === 'right' ? 70 : -70,
+    }),
+    [swipeDirection],
+  );
+  return {
+    setSwipeDirection,
+    exitAnimation,
+    initialAnimation,
+  };
+};
+
+export const Drawer = (React.forwardRef<HTMLDivElement, Props>(
+  ({ className, children, disableContentStyle, onClose, open: isOpenExternal, title }, ref) => {
+    const isControlled = isBoolean(isOpenExternal);
+    const escapePress = useKeyPress('Escape');
+    const [isOpenInternal, setIsOpenInternal] = useState(true);
+    const isOpen = isControlled ? isOpenExternal : isOpenInternal;
+    const isDesktop = useMedia(t => t.media.greaterThan(t.breakpoints.sm)) || false;
+    const seed = useUIDSeed();
+    const uid = seed(displayName);
+    const dragControls = useDragControls();
+
+    const {
+      setSwipeDirection,
+      exitAnimation,
+      initialAnimation,
+    } = useDifferentAnimationBasedOnSwipeDirection();
+
+    const startDrag = useCallback(
+      event => {
+        dragControls.start(event, {
+          snapToCursor: false,
+        });
+      },
+      [dragControls],
+    );
+
+    const handleCloseClick = useCallback(() => {
+      setIsOpenInternal(false);
+
+      if (onClose) {
+        onClose();
+      }
+    }, [onClose]);
+
+    const handleDragEnd = useCallback(
+      (_, info) => {
+        if (info.offset.x < 0) {
+          setSwipeDirection('left');
+        } else {
+          setSwipeDirection('right');
+        }
+        if (shouldCloseBecauseOfDrag(info)) {
+          handleCloseClick();
+        } else {
+          // Todo: handle small offsets -> rollback
+        }
+      },
+      [handleCloseClick],
+    );
+
+    useEffect(() => {
+      if (isOpen && escapePress) {
+        handleCloseClick();
+      }
+    }, [escapePress, handleCloseClick, isOpen]);
+
+    return (
+      <AnimatePresence>
+        {isOpen ? (
+          <Portal>
+            <FocusLock disabled={isDesktop}>
+              <RemoveScroll enabled={!isDesktop}>
+                <Container
+                  className={className}
+                  aria-labelledby={uid}
+                  {...animationProps}
+                  initial={initialAnimation}
+                  exit={exitAnimation}
+                  ref={ref}
+                  dragControls={dragControls}
+                  dragListener={false}
+                  drag="x"
+                  onDragEnd={handleDragEnd}
+                >
+                  <TitleWrapper onTouchStart={startDrag}>
+                    {title && <Title title={title} uid={uid} />}
+                    <CloseButton type="button" variant="neutral" onClick={handleCloseClick}>
+                      <Icon.CrossThin size={CROSS_SIZE} title="Close this drawer" />
+                    </CloseButton>
+                  </TitleWrapper>
+                  {disableContentStyle ? children : <Content>{children}</Content>}
+                </Container>
+              </RemoveScroll>
+            </FocusLock>
+          </Portal>
+        ) : null}
+      </AnimatePresence>
+    );
+  },
+) as any) as React.FC<Props> & {
   components: typeof components;
-} = ({ className, children, onClose, open: isOpenExternal, title }) => {
-  const isControlled = isBoolean(isOpenExternal);
-  const escapePress = useKeyPress('Escape');
-  const [isOpenInternal, setIsOpenInternal] = useState(true);
-  const isOpen = isControlled ? isOpenExternal : isOpenInternal;
-  const isDesktop = useMedia(t => t.media.greaterThan(t.breakpoints.sm)) || false;
-  const seed = useUIDSeed();
-  const uid = seed(displayName);
-
-  const handleCloseClick = useCallback(() => {
-    setIsOpenInternal(false);
-
-    if (onClose) {
-      onClose();
-    }
-  }, [onClose]);
-
-  useEffect(() => {
-    if (isOpen && escapePress) {
-      handleCloseClick();
-    }
-  }, [escapePress, handleCloseClick, isOpen]);
-
-  return isOpen ? (
-    <Portal>
-      <FocusLock disabled={isDesktop}>
-        <RemoveScroll enabled={!isDesktop}>
-          <Container className={className} aria-labelledby={uid} {...animationProps}>
-            <Box mb={2}>
-              <Flexbox container alignItems="baseline">
-                {title && <Title title={title} uid={uid} />}
-                <CloseButton type="button" variant="neutral" onClick={handleCloseClick}>
-                  <Icon.CrossThin size={5} title="Close this drawer" />
-                </CloseButton>
-              </Flexbox>
-            </Box>
-            <FadedScroll>{children}</FadedScroll>
-          </Container>
-        </RemoveScroll>
-      </FocusLock>
-    </Portal>
-  ) : null;
 };
 
 Drawer.components = components;