+++ src/Molecules/Drawer/Drawer.types.ts
@@ -1,13 +1,14 @@
 export type Props = {
   className?: string;
+  /** Useful when integrating with FadedScroll component */
+  disableContentStyle?: boolean;
   onClose?: Function;
   title?: React.ReactNode;
   open?: boolean;
+  ref?: React.Ref<HTMLDivElement>;
 };
 
 export type TitleProps = {
   title: React.ReactNode;
   uid: string;
 };
-
-export type DrawerComponent = React.FC<Props>;